### 2023-11-21 #6080 @Kevin-Mattheus-Moerman
Running `cloc` on the  main branch gives me:
```
Language                     files          blank        comment           code
Python                          25            678            833           1420
Markdown                         4             63              0            182
reStructuredText                 5             34             21             61
YAML                             3             10             22             51
INI                              1              7              0             28
DOS Batch                        1              8              1             26
make                             1              4              7              9
SUM:                            40            804            884           1777
```
