+++ paper.bib
@@ -117,7 +117,7 @@ volume = {62},
 pages = {327-350},
 year = {2014},
 issn = {1364-8152},
-doi = {https://doi.org/10.1016/j.envsoft.2014.07.009},
+doi = {10.1016/j.envsoft.2014.07.009},
 url = {https://www.sciencedirect.com/science/article/pii/S1364815214002102},
 author = {Dean P. Holzworth and Neil I. Huth and Peter G. deVoil and Eric J. Zurcher and Neville I. Herrmann and Greg McLean and Karine Chenu and Erik J. {van Oosterom} and Val Snow and Chris Murphy and Andrew D. Moore and Hamish Brown and Jeremy P.M. Whish and Shaun Verrall and Justin Fainges and Lindsay W. Bell and Allan S. Peake and Perry L. Poulton and Zvi Hochman and Peter J. Thorburn and Donald S. Gaydon and Neal P. Dalgliesh and Daniel Rodriguez and Howard Cox and Scott Chapman and Alastair Doherty and Edmar Teixeira and Joanna Sharp and Rogerio Cichota and Iris Vogeler and Frank Y. Li and Enli Wang and Graeme L. Hammer and Michael J. Robertson and John P. Dimes and Anthony M. Whitbread and James Hunt and Harm {van Rees} and Tim McClelland and Peter S. Carberry and John N.G. Hargreaves and Neil MacLeod and Cam McDonald and Justin Harsdorf and Sara Wedgwood and Brian A. Keating},
 keywords = {APSIM, Simulation, Model, Agricultural systems, Crop, Farming system, Gene-to-phenotype model},
