### 2023-12-05 #6080 @jucordero
@Kevin-Mattheus-Moerman Thanks!
Quick question: Is it okay to modify the .md for the paper at this stage? Some coauthors have provided updated ORCID numbers and affiliations.
