### 2023-12-05 #6080 @Kevin-Mattheus-Moerman
Yes, that is absolutely fine. You can call `@editorialbot generate pdf` here in a comment to update the paper draft then as well.
