### 2023-12-07 b4493aa
Adjusted some refs, author info
