+++ paper.bib
@@ -1,7 +1,7 @@
 @article{
   PN18,
   author = {J. Poore  and T. Nemecek },
-  title = {Reducing food\&\#x2019;s environmental impacts through producers and consumers},
+  title = {Reducing food's environmental impacts through producers and consumers},
   journal = {Science},
   volume = {360},
   number = {6392},
@@ -35,7 +35,7 @@
 @misc{CEH,
         doi = {10.5285/d6f8c045-521b-476e-b0d6-b3b97715c138},
         url = {https://doi.org/10.5285/d6f8c045-521b-476e-b0d6-b3b97715c138},
-     author = {Morton, R.D.;Marston, C.G.;O’Neil, A.W.;Rowland, C.S.},
+     author = {Morton, R.D.;Marston, C.G.;O'Neil, A.W.;Rowland, C.S.},
   publisher = {NERC EDS Environmental Information Data Centre},
       title = {Land Cover Map 2020 (1km summary rasters, GB and N. Ireland)},
        year = {2022}
@@ -111,7 +111,7 @@ DOI = {10.5194/gmd-14-3007-2021}
 }
 
 @article{APSIM,
-title = {APSIM – Evolution towards a new generation of agricultural systems simulation},
+title = {APSIM - Evolution towards a new generation of agricultural systems simulation},
 journal = {Environmental Modelling & Software},
 volume = {62},
 pages = {327-350},
