### 2023-12-07 bd6d7d5
Fixed bob's author info
