### 2024-01-29 #6305 @editorialbot
[REVIEW]: AgriFoodPy: a package for modelling food systems
**Submitting author:** @jucordero (<a href="http://orcid.org/0000-0002-6625-7656">Juan P. Cordero</a>)
**Repository:** https://github.com/FixOurFood/AgriFoodPy/
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.1.1
