### 2024-02-14 e030133
Plot tests, fixed plot and SSR/IDR functionality
