+++ agrifoodpy/food/tests/test_food.py
@@ -1,6 +1,6 @@
 import numpy as np
 import xarray as xr
-from agrifoodpy.food.food import FoodBalanceSheet
+from agrifoodpy.food.food import FoodBalanceSheet, FoodElementSheet
 from agrifoodpy.food.food import FoodSupply
 import pytest
 
@@ -152,7 +152,8 @@ def test_SSR():
         data_vars=dict(
             imports=(["Year", "Item"], [[10, 20], [30, 40]]),
             exports=(["Year", "Item"], [[5, 10], [15, 20]]),
-            production=(["Year", "Item"], [[50, 60], [70, 80]])
+            production=(["Year", "Item"], [[50, 60], [70, 80]]),
+            domestic=(["Year", "Item"], [[55, 70], [85, 100]])
             ),
 
     coords=dict(Item=("Item", items), Year=("Year", years))
@@ -161,7 +162,6 @@ def test_SSR():
     fbs = FoodBalanceSheet(ds)
 
     # Test basic result on all items
-
     result_basic = fbs.SSR()
     ex_result_basic = xr.DataArray([0.88, 0.810810], dims=("Year"),
                                 coords={"Year": years})
@@ -183,6 +183,13 @@ def test_SSR():
 
     xr.testing.assert_allclose(result_peritem, ex_result_peritem)
 
+    # Test with domestic use
+    result_domestic = fbs.SSR(domestic="domestic")
+    ex_result_domestic = xr.DataArray([0.88, 0.810810], dims=("Year"),
+                                      coords={"Year": years})
+    
+    xr.testing.assert_allclose(result_domestic, ex_result_domestic)
+
 def test_IDR():
 
     items = ["Beef", "Apples"]
@@ -192,7 +199,8 @@ def test_IDR():
         data_vars=dict(
             imports=(["Year", "Item"], [[10, 20], [30, 40]]),
             exports=(["Year", "Item"], [[5, 10], [15, 20]]),
-            production=(["Year", "Item"], [[50, 60], [70, 80]])
+            production=(["Year", "Item"], [[50, 60], [70, 80]]),
+            domestic=(["Year", "Item"], [[55, 70], [85, 100]]),
             ),
       
     coords=dict(Item=("Item", items), Year=("Year", years))
@@ -221,6 +229,13 @@ def test_IDR():
                                      coords={"Year": years, "Item": items})
     
     xr.testing.assert_allclose(result_peritem, ex_result_peritem)
+
+    # Test with domestic use
+    result_domestic = fbs.IDR(domestic="domestic")
+    ex_result_domestic = xr.DataArray([0.24, 0.37837838], dims=("Year"),
+                                      coords={"Year": years})
+    
+    xr.testing.assert_allclose(result_domestic, ex_result_domestic)
 
 def test_scale_add():
 
@@ -365,3 +380,91 @@ def test_scale_element():
     result_year = fbs.scale_element("production", scale_year)
     xr.testing.assert_equal(result_year["production"],
                             [[0.5], [2.0]]*ds["production"])
+
+def test_plot_bars():
+
+    items = ["Beef", "Apples"]
+    years = [2020, 2021]
+
+    ds = xr.Dataset(
+        data_vars=dict(
+            imports=(["Year", "Item"], [[10, 20], [30, 40]]),
+            exports=(["Year", "Item"], [[5, 10], [15, 20]]),
+            production=(["Year", "Item"], [[50, 60], [70, 80]])
+            ),
+        coords=dict(Item=("Item", items), Year=("Year", years)))
+    
+    fbs = FoodBalanceSheet(ds)
+
+    # Test default call
+    ax_default = fbs.plot_bars()
+
+    assert len(ax_default.patches) == 6
[…]
