### 2024-02-21 9abc5f0
Plot tests, fixed plot and SSR/IDR functionality (#60)
* Plot tests, fixed plot and SSR/IDR functionality

* land plot test
