+++ agrifoodpy/food/food.py
@@ -11,6 +11,7 @@ FoodBalanceSheet or FoodElementSheet style xarray primitive.
 import numpy as np
 import xarray as xr
 import copy
+import warnings
 
 from agrifoodpy.array_accessor import XarrayAccessorBase
 
@@ -206,8 +207,8 @@ class FoodBalanceSheet(XarrayAccessorBase):
         
         return out
 
-    def SSR(self, items=None, per_item=False, production="production",
-            imports="imports", exports="exports"):
+    def SSR(self, items=None, per_item=False, domestic=None,
+            production="production", imports="imports", exports="exports"):
         """Self-sufficiency ratio
 
         Self-sufficiency ratio (SSR) or ratios for a list of item imports,
@@ -223,6 +224,8 @@ class FoodBalanceSheet(XarrayAccessorBase):
             list is provided, the SSR is computed for all items.
         per_item : bool, optional
             Whether to return an SSR for each item separately. Default is false
+        domestic : string, optional
+            Name of the DataArray containing the domestic use data
         production : string, optional
             Name of the DataArray containing the production data
         imports : string, optional
@@ -245,12 +248,18 @@ class FoodBalanceSheet(XarrayAccessorBase):
                 items = [items]
             fbs = fbs.sel(Item=items)
 
+        if domestic is not None:
+            domestic_use = fbs[domestic]
+        else:
+            domestic_use = fbs[production] + fbs[imports] - fbs[exports]
+
         if per_item:
-            return fbs[production] / (fbs[production] + fbs[imports] - fbs[exports])
+            return fbs[production] / domestic_use
 
-        return fbs[production].sum(dim="Item") / (fbs[production]+fbs[imports]-fbs[exports]).sum(dim="Item")
+        return fbs[production].sum(dim="Item") / domestic_use.sum(dim="Item")
 
-    def IDR(self, items=None, per_item=False):
+    def IDR(self, items=None, per_item=False, imports="imports", domestic=None,
+            production="production", exports="exports"):
         """Import-dependency ratio
 
         Import-ependency ratio (IDR) or ratios for a list of item imports,
@@ -266,6 +275,15 @@ class FoodBalanceSheet(XarrayAccessorBase):
             list is provided, the IDR is computed for all items.
         per_item : bool, optional
             Whether to return an IDR for each item separately. Default is false.
+        domestic : string, optional
+            Name of the DataArray containing the domestic use data
+        imports : string, optional
+            Name of the DataArray containing the imports data
+        exports : string, optional
+            Name of the DataArray containing the exports data
+        production : string, optional
+            Name of the DataArray containing the production data
+            
 
         Returns
         -------
@@ -282,10 +300,15 @@ class FoodBalanceSheet(XarrayAccessorBase):
                 items = [items]
             fbs = fbs.sel(Item=items)
 
-        if per_item:
-            return fbs["imports"] / (fbs["production"] + fbs["imports"] - fbs["exports"])
+        if domestic is not None:
+            domestic_use = fbs[domestic]
+        else:
+            domestic_use = fbs[production] + fbs[imports] - fbs[exports]
 
-        return fbs["imports"].sum(dim="Item") / (fbs["production"]+fbs["imports"]-fbs["exports"]).sum(dim="Item")
+        if per_item:
+            return fbs["imports"] / domestic_use
+
+        return fbs["imports"].sum(dim="Item") / domestic_use.sum(dim="Item")
 
     def plot_bars(self, show="Item", elements=None, inverted_elements=None,
                   ax=None, colors=None, labels=None, **kwargs):
[…]
