### 2024-02-28 #6305 @jsun
Hi, @jucordero, @crvernon, I have not yet finished checking all the items; but I have written below with my comments on the parts I have checked so far.

**Documentation**

- The documentation (probably README.md) gave several examples, however, I think more documentation is required for this package. For example, what kind of functions/methods are implemented in this package, what are the arguments of the function, ...

**Software paper**

-  It is recommended to write a more concise overview of the software in **Summary** section. Please consider writing about detailed features of the software and features to be implemented in the future in **Features** section.
- The background was well-written, but more emphasis should be placed on what exactly this software can do and what it can do better than existing software.
- The structure of the paper needs proofreading.

Additional comments

- In the paper, the authors write a lot about future releases and implementations. This underlines that the development of the package is incomplete. I suggest that author implement these features during this review process.
