### 2024-03-04 #6305 @crvernon
@jucordero yes, no problem at all to update the main branch during review as long as those changes have been made known to reviewers in this thread.  These changes should also not interfere with any current features that the reviewers are currently reviewing.  So you may want to check with them directly as well.  Anytime you link to this thread in your pull requests or issues it will show up here so all can track progress.
