### 2024-03-27 #6305 @jucordero
Hi @crvernon

@jsun has made a few suggestions on the documentation and software paper which we have been addressing.
Simultaneously we have been working on one of the main functionalities which was missing in the first release, but haven't discussed with the reviewers yet on whether it is fine to merge this right now.

I haven't heard from @kanishkan91

Happy Easter!
