### 2024-04-05 FixOurFood/AgriFoodPy#66 @kanishkan91
JOSS review: No class named "FAOSTAT" is available as mentioned in example or the same does not work with current instructions
The example in the README is based on a FAOSTAT class that does not exist. Should the example be based on existing available classes such as  "FoodBalanceSheet" ?

It seems the authors should add the "pip install git+https://github.com/FixOurFood/agrifoodpy-data.git@importable" instructions explicitly in the README for the example to work correctly

https://github.com/openjournals/joss-reviews/issues/6305
