### 2024-04-05 FixOurFood/AgriFoodPy#67 @kanishkan91
JOSS review: fbs.plot_years() from README does not work
Not sure what is going on but I got this error (attached below). If this is because I'm missing some packages, those should be added to the dependencies.

![image](https://github.com/FixOurFood/AgriFoodPy/assets/37234086/be9dad50-dc98-474b-80dc-9d39f404042b)

https://github.com/openjournals/joss-reviews/issues/6305
