### 2024-04-05 FixOurFood/AgriFoodPy#68 @kanishkan91
JOSS review: Need some edits for README examples to work correctly
I think two lines need to be added for the plots in the README to actually show.

First, you would need this at the imports
`import matplotlib.pyplot as plt`

Secondly, you would need this line at the very end for plots to actually show-
`plt.show()`

Otherwise you only get an axes object that never actually plots.

https://github.com/openjournals/joss-reviews/issues/6305
