### 2024-04-05 FixOurFood/AgriFoodPy#69 @kanishkan91
JOSS review: Need link to the full docs for all classes and methods available.
Unfortunately, there are several really useful classes and methods in this package that are currently not documented.

I see that the authors "built the docs" for the same using sphinx.

Can the authors add a link to the full docs somewhere. Can the sphinx documentation be linked in the "About" section?

https://github.com/openjournals/joss-reviews/issues/6305
