### 2024-04-05 FixOurFood/AgriFoodPy#70 @kanishkan91
JOSS review: Error when using reforestation example
The examples available on this package are really useful!

Though I ran into an issue when it comes to the reforestation example here- https://github.com/FixOurFood/AgriFoodPy/blob/main/examples/modules/plot_reforest_ALC_4_5.py

I think there's something really simple that goes wrong with respect to the classes on this line-

`ALC.land.plot(ax=axes[1])`

https://github.com/openjournals/joss-reviews/issues/6305
