### 2024-04-05 FixOurFood/AgriFoodPy#71 @kanishkan91
JOSS review: Please add link somewhere to the actual examples that are available on the repo
The authors have added really useful examples here- https://github.com/FixOurFood/AgriFoodPy/tree/main/examples/modules

However, users have no way of knowing there are some detailed examples available already. I would recommend that the authors add a link somewhere in the README to these examples, so the users would know that they exist.

Else this gives the impression that no examples are available, which is clearly untrue.

https://github.com/openjournals/joss-reviews/issues/6305
