### 2024-04-05 FixOurFood/AgriFoodPy#72 @kanishkan91
JOSS review: Minor comment but should there be version requirements for dependencies? 
For the depoendencies documented here - https://github.com/FixOurFood/AgriFoodPy/blob/main/environment.yml,

should there be version requirements?

https://github.com/openjournals/joss-reviews/issues/6305
