+++ tox.ini
@@ -1,5 +1,7 @@
 [tox]
-requires = tox-conda
+requires =
+    tox-conda
+    setuptools
 envlist =
 	py{39,310}-test{,-latest,-oldest}
     build_docs
