+++ .github/workflows/test-conda.yml
@@ -1,4 +1,4 @@
-name: Python Package using Conda
+name: Tests
 
 on: [push, pull_request]
 
