### 2024-04-30 FixOurFood/AgriFoodPy#67 @jucordero
Hi @kanishkan91
Thank you very much for your time on reviewing AgriFoodPy.

I've looked into this and there seems to be an [issue](https://github.com/matplotlib/matplotlib/issues/26463) with PyCharm rather than a missing AgriFoodPy dependency.
According to this issue, the problem should be resolved by updating PyCharm to a more recent version (You are running 2021.3.3 which is somewhat old)
