### 2024-04-30 13187b4
Added explicit plot.show() to readme
