### 2024-04-30 7215493
Added instructions to install afp_data
