### 2024-05-01 #6305 @jucordero
Hi @crvernon, @jsun and @kanishkan91
Once again, thank you very much for your comments and suggestions. These have helped us improve the accessibility to the existing documentation, the examples and the quality of the experience for first users when landing on the package repository.
Apologies for the delay of this response. It has been a few very busy weeks lately.

@jsun
**Documentation**
Documentation for the package in the form of an auto-API document already exists for the package. It is automatically generated and available in readthedocs. We have included links to it both in the README and the about section of the repository to make it more visible. These have been addressed in https://github.com/FixOurFood/AgriFoodPy/pull/75

@kanishkan91
- In [PR#74](https://github.com/FixOurFood/AgriFoodPy/pull/74) we have fixed the issue with the reforestation example not running properly. This was caused by an object imported from the agrifoodpy_data package, which changed from a DataArray to a Dataset. This fixes [Issue 70](https://github.com/FixOurFood/AgriFoodPy/issues/70).
- In [PR#75](https://github.com/FixOurFood/AgriFoodPy/pull/75) we made several changes to the README to address some of the issues you have raised. We added a new section to point to the documentation ([Issue 69](https://github.com/FixOurFood/AgriFoodPy/issues/69)) and examples ([Issue 71](https://github.com/FixOurFood/AgriFoodPy/issues/71)). As mentioned above we also added a link to the readthedocs documentation in the about section of the repository.
- Made small changes to the code snippet in the README, including explicit instructions to install the optional agrifoodpy_data package and statements to correctly display the generated plots. This addresses [Issue 66](https://github.com/FixOurFood/AgriFoodPy/issues/66) and [Issue 68](https://github.com/FixOurFood/AgriFoodPy/issues/68)

There are still a few outstanding tasks which we haven't completed but are still working on.

- Edits to the software paper suggested by @jsun: We are in the process of editing the summary and adding a "Functionality" section to the paper. This should be finished early next week.
- Version requirements for the dependencies, raised by @kanishkan91: I'm not entirely sure on what criteria to use to define the minimum required versions for each dependency. Still investigating this.

As soon as we complete those tasks I'll ping you in this thread.
