### 2024-05-01 6b7119b
JOSS README review (#75)
* Documentation badge

* Tests badge

* Added explicit plot.show() to readme

* Added instructions to install afp_data

* Update README.md
