+++ README.md
@@ -1,5 +1,8 @@
 # AgriFoodPy
 
+[![Documentation Status](https://readthedocs.org/projects/agrifoodpy/badge/?version=latest)](https://agrifoodpy.readthedocs.io/en/latest/?badge=latest)
+[![Tests](https://github.com/FixOurFood/AgriFoodPy/actions/workflows/test-conda.yml/badge.svg)](https://github.com/FixOurFood/AgriFoodPy/actions/workflows/test-conda.yml)
+
 AgriFoodPy is a collection of methods for manipulating and modelling agrifood
 data. It provides modelling for a variety of aspects of the food system,
 including food consumption paterns, environmental impact and emissions data,
@@ -24,6 +27,13 @@ AgriFoodPy can be installed using _pip_, by running
 pip install agrifoodpy
 ```
 
+UK data to test the package is available from the agrifoodpy_data repository
+which currently can be installed using
+
+```bash
+pip install git+https://github.com/FixOurFood/agrifoodpy-data.git@importable
+```
+
 ## Usage:
 
 Each of the four basic modules on AgriFoodPy (Food, Land, Impact, Population)
@@ -31,16 +41,18 @@ has its own set of basic array manipulation functionality, a set of
 modelling methods to extract basic metrics from datasets, and interfaces with
 external modelling packages and code.
 
-Agrifoodpy employs _xarray_ accesors to add additional functionality on top of
-the array manipulation
+Agrifoodpy employs _xarray_ accesors to provide additional functionality on top
+of the array manipulation provided by xarray.
 
-Basic usage of the accesors depend on the type of array being manipulated
-As an example, using the food module
+Basic usage of the accesors depend on the type of array being manipulated.
+The following examples uses the **food** module with the importable UK data
+mentioned above: 
 
 ```python
 # import the FoodBalanceSheet accessor and FAOSTAT from agrifoodpy_data
 from agrifoodpy.food.food import FoodBalanceSheet
 from agrifoodpy_data.food import FAOSTAT
+import matplotlib.pyplot as plt
 
 # Extract data for the UK (Region=229)
 food_uk = FAOSTAT.sel(Region=229)
@@ -50,6 +62,7 @@ SSR = food_uk.fbs.SSR(per_item=True)
 
 # Plot the results using the fbs accessor plot_years function
 SSR.fbs.plot_years()
+plt.show()
 ```
 
 To use the specific models and interfaces to external code, these need to be
@@ -60,6 +73,7 @@ imported
 from agrifoodpy.food.food import FoodBalanceSheet
 from agrifoodpy_data.food import FAOSTAT
 import agrifoodpy.food.model as food_model
+import matplotlib.pyplot as plt
 
 # Extract data for the UK in 2020 (Region=229, Year=2020)
 food_uk = FAOSTAT.sel(Region=229, Year=2020)
@@ -75,11 +89,20 @@ food_uk_scaled = food_model.balanced_scaling(food_uk,
 # Plot bar summary of resultant food quantities
 food_uk_scaled.fbs.plot_bars(elements=["production","imports"],
                             inverted_elements=["exports","food"])
+plt.show()
 ```
 
 In he future, we plan to implement a pipeline manager to automatize certain
 aspects of the agrifood execution, and to simulate a comprehensive model where
 all aspects of the food system are considered simultaneously.
+
+## Examples and documentation
+
+[Examples](https://agrifoodpy.readthedocs.io/en/latest/examples/index.html#modules)
+demonstrating the functionality of AgriFoodPy can be the found in the
+[package documentation](https://agrifoodpy.readthedocs.io/en/latest/).
+These include the use of accessors to manipulate data and access to basic
+models.
 
 ## Contributing
 
