+++ paper.bib
@@ -46,7 +46,7 @@
   title     = {United Nations, Department of Economic and Social
               Affairs, Population Division (2022). World Population Prospects
               2022: Summary of Results. UN DESA/POP/2022/TR/NO. 3.},
-  author    = {"United Nations"},
+  author    = "{United Nations}",
   year      = {2022}
 }
 
