### 2024-05-15 f51d4b2
Dependencies version requirements
