+++ .github/workflows/test-conda.yml
@@ -20,6 +20,7 @@ jobs:
         echo $CONDA/bin >> $GITHUB_PATH
     - name: Install dependencies
       run: |
+        conda install -y python=3.9
         conda env update --file environment.yml --name base
     - name: Test with tox
       run: |
