+++ environment.yml
@@ -4,7 +4,7 @@ channels:
   - defaults
 
 dependencies:
-  - python>=3.10
+  - python>=3.9
   - pip>=24.0
   - numpy>=1.26.4
   - netcdf4>=1.6.2
