### 2024-05-16 d97ae34
Update GitHub actions (#77)
* Updated git actions

* Updated git actions

* versions for dependencies

* Python version
