### 2024-05-16 #6305 @jucordero
@jsun
I have updated the manuscript.
- Made a few changes to the summary to make it more concise.
- Added a features section. Also rewrote a few paragraphs to remove some redundant information
- Fixed some typos and changed to a more consistent use of concepts.

@kanishkan91
In [PR#77](https://github.com/FixOurFood/AgriFoodPy/pull/77) I have added version requirements to the modules to set the conda environment.
Also took the opportunity to update deprecated GitHub actions which were causing issues with the unit tests.
