### 2024-05-17 #6305 @crvernon
:wave: @jucordero - I am conducting my pass now and have a few comments that need to be addressed before moving forward:

In the paper:

- [ ] LINES 24-25: The statement "...layer between for data sources..."  seems off.  Should the "for" be removed?
- [ ] LINE 59:  The statement "....the importance of food production on climate change." reads a bit odd.  How about "...the importance of accounting for the impacts of food production on climate change." or something like that.
- [ ] LINE 75: The word "The" does not need to be capitalized in "...The Environmental..."
- [ ] There is an issue with the formatting of your Gillies et al. paper which has two hyphens after year 2013.  This will likely need to be corrected in your bib file.
- [ ] Please check to see if the UN 2022 document you site has a DOI or weblink available.

Once you address these I'll provide further instruction on how we can move forward with getting this submission ready for acceptance.  Thanks!
