### 2024-05-17 #6305 @jucordero
Thanks for your comments, @crvernon.

I have made the following changes to the paper

- Removed stray 'for' in line 24-25
- In line 59, opted for a more straightforward wording. Used "have reported the impact
of food production on climate change" instead.
- Removed capitalization of the word 'The' in line 75
- Edited the bibtex for the Gilles et al. `Rasterio`  paper. It is a bit odd but that is the official bibtex entry they have in their repository citation guidelines.
- I have changed the reference to the FAOSTAT dataset to point to the latest published report, which does has its corresponding DOI. This is probably the correct way to cite the Food Balance Sheets anyways.
