### 2024-05-17 dab8fc9
Typos, citations, styling of some phrases
