+++ paper.bib
@@ -25,11 +25,11 @@
 }
 
 @article{FAOSTAT,
-  title     = {FAO. FAOSTAT. License: CC BY-NC-SA 3.0 IGO. Extracted from:
-              https://www.fao.org/faostat/en/#data/FBS. Date of Access:
-              22-06-2023},
+  title     = {FAO. 2023. Food balance sheets 2010–2021. Global, regional and country trends. FAOSTAT AnalyticalBrief Series No. 72.},
   author    = {FAO},
-  year      = {2022}
+  year      = {2023},
+  url       = {https://doi.org/10.4060/cc8088en},
+  doi       = {10.4060/cc8088en}
 }
 
 @misc{CEH,
@@ -94,7 +94,7 @@
   author =    {Sean Gillies and others},
   organization = {Mapbox},
   title =     {Rasterio: geospatial raster I/O for {Python} programmers},
-  year =      {2013--},
+  year =      {2013},
   url = "https://github.com/rasterio/rasterio"
 }
 
