### 2024-05-18 #6305 @crvernon
@jucordero thank you!  Concerning the following:
>I have changed the reference to the FAOSTAT dataset to point to the latest published report, which does have a corresponding DOI. This is probably the correct way to cite the Food Balance Sheets anyways.

I still see the UN 2022 reference in the paper.  Did you mean to change it out elsewhere as well?
