### 2024-05-20 #6305 @crvernon
@jucordero no problem.  Yes, I think including that link is better than not having one at all if it matches the citation standards.  Just let me know when it is done so we can move forward.  Thanks!
