+++ paper.bib
@@ -47,7 +47,8 @@
               Affairs, Population Division (2022). World Population Prospects
               2022: Summary of Results. UN DESA/POP/2022/TR/NO. 3.},
   author    = "{United Nations}",
-  year      = {2022}
+  year      = {2022},
+  url       = {https://population.un.org/wpp/},
 }
 
 @software{geopandas,
