### 2024-05-21 #6305 @crvernon
👋  @jucordero - we are almost there! Next is just setting up the archive for your new release.

So here is what we have left to do:

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list matches the people and order in your paper). You may also add the authors' ORCID.

- [ ]  Please respond with the 1) version and 2) DOI of the archived version here
