### 2024-05-21 #6305 @jucordero
@crvernon
I have just created Release v0.1.2 and its corresponding entry in Zenodo

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.11236802.svg)](https://doi.org/10.5281/zenodo.11236802)
