### 2024-05-25 #6305 @Kevin-Mattheus-Moerman
@jucordero as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Most seems in order, I only have the below point that require your attention:

- [ ] In the affiliations, please spell out `UK` as `United Kingdom`
